worked for implementing final functions for the telegram bot.
first one is history of search, that being stored using sqlite locally on pc.

second one is search by budget with two options: high budget and low budget movies.

also added some visual aesthetic to the bot.

this is final version of the bot with the most functions working.

some of the files are being used for my own reference to make development easier.